+++ scilab/modules/core/sci_gateway/c/sci_getscilabmode.c
@@ -36,7 +36,7 @@ int C2F(sci_getscilabmode) _PARAMS((char *fname,unsigned long fname_len))
 	}
 
 	n1=1;
-	CreateVarFromPtr(Rhs+1,STRING_DATATYPE,(m1=strlen(output), &m1),&n1,&output);
+	CreateVarFromPtr(Rhs+1,STRING_DATATYPE,(m1=(int)strlen(output), &m1),&n1,&output);
 	if (output) {FREE(output);output=NULL;}
 
 	LhsVar(1) = Rhs+1;
