+++ scilab/modules/core/src/c/cluni0.c
@@ -43,7 +43,7 @@ int C2F(cluni0)(char *in_name, char *out_name, int *out_n, long int lin, long in
 		  strncpy(out_name,in_name,(size_t)lin);
 		  out_name[lin]='\0';
       }
-  *out_n = strlen(out_name);
+  *out_n = (int)strlen(out_name);
 #ifdef _MSC_VER
   for (k=0 ; k < *out_n ;k++) if ( out_name[k]=='/') out_name[k]='\\';
 #else
