+++ scilab/modules/gui/sci_gateway/c/sci_ClipBoard.c
@@ -37,7 +37,7 @@ int C2F(sci_ClipBoard) _PARAMS((char *fname,unsigned long l))
 
 				if (output)
 				{
-					CreateVarFromPtr(Rhs+ 1,STRING_DATATYPE,(m1=strlen(output), &m1),&n1,&output);
+					CreateVarFromPtr(Rhs+ 1,STRING_DATATYPE,(m1=(int)strlen(output), &m1),&n1,&output);
 					FREE(output);
 					LhsVar(1)=Rhs+1;
 				}
@@ -127,7 +127,7 @@ int C2F(sci_ClipBoard) _PARAMS((char *fname,unsigned long l))
 
 						for (i=0; i<m1; i++) for (j=0; j<n1; j++) 
 						{
-							SizeofTextToSendInClipboard=SizeofTextToSendInClipboard+strlen(Str[j*m1+i])+strlen(" \n");
+							SizeofTextToSendInClipboard=SizeofTextToSendInClipboard+(int)strlen(Str[j*m1+i])+(int)strlen(" \n");
 							buffer[i*n1+j]=(char*)MALLOC(strlen(Str[j*m1+i])+1);
 							sprintf(buffer[i*n1+j],"%s",Str[j*m1+i]);
 						}
